refactor: DB에 저장된 데이터를 통해 오목 게임을 복구시키는 기능을 별도의 클래스로 분리
