+++ app/src/main/java/woowacourse/omok/omokdb/OmokRestoreData.kt
@@ -1,0 +1,53 @@
+package woowacourse.omok.omokdb
+
+import android.widget.ImageView
+import android.widget.TableLayout
+import android.widget.TableRow
+import androidx.core.view.children
+import omok.model.position.Position
+import omok.model.stone.BlackStone
+import omok.model.stone.GoStone
+import omok.model.stone.WhiteStone
+
+object OmokRestoreData {
+    private const val BLACK_STONE_VALUE = "흑"
+
+    fun restoreGameData(
+        dbHelper: OmokDbHelper,
+        board: TableLayout,
+        imageView: (GoStone) -> Int,
+    ) {
+        dbHelper.selectStonesInfo().forEach { (position, stoneColor) ->
+            recoverBoard(board, position, stoneColor, imageView)
+        }
+    }
+
+    private fun recoverBoard(
+        board: TableLayout,
+        position: Position,
+        stoneColor: String,
+        imageView: (GoStone) -> Int,
+    ) {
+        board.children
+            .filterIsInstance<TableRow>()
+            .flatMap { it.children }
+            .filterIsInstance<ImageView>()
+            .forEachIndexed { index, view ->
+                if (Position.fromIndex(index) == position) {
+                    recoverStone(stoneColor.stone(), position, view, imageView)
+                }
+            }
+    }
+
+    private fun recoverStone(
+        stone: GoStone,
+        position: Position,
+        view: ImageView,
+        imageView: (GoStone) -> Int,
+    ) {
+        view.setImageResource(imageView(stone))
+        stone.putStone(position)
+    }
+
+    private fun String.stone() = if (this == BLACK_STONE_VALUE) BlackStone else WhiteStone
+}
